backend/pyproject.toml
[project]
name = "humanoid-robotics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "cohere>=5.20.0",
    "fastapi[standard]>=0.124.4",
    "openai>=2.12.0",
    "openai-agents>=0.6.3",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "trafilatura>=2.0.0",
    "typer>=0.20.0",
    "uv>=0.9.17",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["backend"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.9",
]
